requirements.txt
feedparser==6.0.11
anthropic>=0.76.0
python-telegram-bot==21.7
python-dotenv==1.0.1
schedule==1.2.2
requests==2.32.3

# Phase 1: Production-ready improvements
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
